feat: move tests to test folder
