+++ tests/loaders/test_fileSavers.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import mock_open, patch
 
-from .fileSavers import FileSavers
+from ...methods.loaders.fileSavers import FileSavers
 
 
 class FileSaversTestSuite(TestCase):
